pyproject.toml
[build-system]
build-backend = 'hatchling.build'
requires = [ "hatch-fancy-pypi-readme>=22.5", "hatchling" ]

[project]
name = "splitme-ai"
version = "0.1.9"
authors = [ { name = "Eli Salamie", email = "egsalamie@gmail.com" } ]
description = "Break down docs, build up knowledge."
keywords = [
    "ai",
    "devtools",
    "documentation",
    "documentation-tool",
    "markdown",
    "markdown-parser",
    "markdown-split",
    "mkdocs",
    "python",
    "readme",
]
license = "MIT"
readme = "README.md"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Operating System :: POSIX :: Linux",
    "Environment :: Console",
    "Environment :: MacOS X",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet",
    "Framework :: Pydantic",
    "Framework :: Pytest",
]
requires-python = ">=3.9"

dependencies = [
    "aiofiles>=24.1.0",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "pyyaml>=6.0.2",
    "toml>=0.10.2 ; python_full_version < '3.11'",
]

[project.urls]
Homepage = "https://docs.splitme-ai.com"
Source = "https://github.com/eli64s/splitme-ai"
Documentation = "https://docs.splitme-ai.com"
Changelog = "https://github.com/eli64s/splitme-ai/tags"

[project.optional-dependencies]
ai = [
    "pydantic-ai>=0.0.15",
]

[project.scripts]
splitme-ai = "splitme_ai.__main__:run_async_main"

[dependency-groups]
dev = [
    "hatch>=1.13",
]
test = [
    "coverage[toml]>=7.6.9",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6",
    "pytest-pretty>=1.2",
    "pytest-randomly>=3.16",
    "pytest-sugar>=1",
    "pytest-xdist>=3.6.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.47",
]
lint = [
    "mypy>=1.14",
    "pyright>=1.1.391",
    "ruff>=0.8.4",
    "types-aiofiles>=24.1.0.20241221",
    "types-pyyaml>=6.0.12.20240917",
]

[tool.hatch.build.targets.wheel]
packages = [ "src/splitme_ai" ]

# [tool.hatch.version]
# path = 'src/splitme_ai/__init__.py'

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [
    '/README.md',
    '/Makefile',
    '/splitme',
    '.splitme-ai',
    '/tests',
]

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = 'text/markdown'
substitutions = [
    { pattern = '(\s+)#(\d+)', replacement = '\1[#\2](https://github.com/eli64s/splitme-ai/issues/\2)' },
    { pattern = '(\s+)@([\w\-]+)', replacement = '\1[@\2](https://github.com/\2)' },
]

[tool.pytest.ini_options]
addopts = [
    "-vv",
    "-rfEsxX",
    "--tb=short",
    "--durations=3",
    "--cov=src/splitme_ai",
    "--cov-branch",
    "--cov-fail-under=0",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=json:tests/.reports/coverage.json",
    "--asyncio-mode=auto",
    "--numprocesses=auto",
]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def repr",
    "raise NotImplementedError",
    "if name == .__main__.:",
    "if TYPE_CHECKING:",
    "pass",
    "\\.\\.\\.",
]
precision = 2
show_missing = true
skip_empty = true
sort = "cover"

[tool.coverage.run]
data_file = "tests/.reports/.coverage"
omit = [
    "docs/*",
    "examples/*",
    "tests/*",
    "**/__init__.py",
]
source = ["splitme_ai"]

[tool.coverage.json]
output = "tests/.reports/coverage.json"
pretty_print = true

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnnecessaryIsInstance = false
reportUnnecessaryTypeIgnoreComment = true
include = [ "splitme_ai", "tests" ]
venvPath = ".venv"
executionEnvironments = [
    { root = "tests", reportUnusedFunction = false },
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
